backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-redis==6.0.0
psycopg2-binary==2.9.10
djangorestframework-authtoken==1.2.0
django-filter==25.1
